+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `GaussianNB` to `naive_bayes`
 - `Multinomial` and `Normal` to `dist`
 - `PassiveAggressiveI` and `PassiveAggressiveII` to `optim`
+- `OneVsRestClassifier`` to `multiclass`
 
 ### Removed